Code maker program new version Delphi
It was created in order to help every novice programmer in programming, especially in creating applications that deal with databases
It is currently focused on access databases. If circumstances help me, I will develop it, God willing.
And I will put the source code, I mean the script and the application together. In order to be seen by those who wanted Delphi programmers.
I also ask the Delphi programmers to point me to any errors they see in the source code.
Each programmer can develop or modify it. It is not for trading, it is a charitable and free project.